### 2024-03-13 #6477 @editorialbot
[PRE REVIEW]: ahead: Univariate and multivariate time series forecasting with uncertainty quantification (including simulation approaches)'
**Submitting author:** @thierrymoudiki (<a href="http://orcid.org/0000-0002-9422-5459">Thierry Moudiki</a>)
**Repository:** https://github.com/Techtonique/ahead
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.11.0
