### 2024-03-13 #6477 @arfon
@thierrymoudiki  – your paper metadata needs fixing up still. I started fixed it here: https://github.com/Techtonique/ahead/pull/15#pullrequestreview-1935099728

Also, please limit the references in your bibliography to only the ones that you are actually citing in your paper.
