+++ R/fitforecast.R
@@ -1,0 +1,535 @@
+#' Fit and forecast for benchmarking purposes
+#'
+#' @param y A univariate time series of class \code{ts}
+#' @param h Forecasting horizon (default is \code{NULL}, in that case, \code{pct_train}
+#' and \code{pct_calibration} are used)
+#' @param pct_train Percentage of data in the training set, when \code{h} is \code{NULL}
+#' @param pct_calibration Percentage of data in the calibration set for conformal prediction
+#' @param method For now "thetaf" (default), "arima", "ets", "tbats", "tslm", "dynrmf" (from ahead),
+#' "ridge2f" (from ahead), "naive", "snaive"
+#' @param level Confidence level for prediction intervals in %, default is 95
+#' @param B Number of bootstrap replications or number of simulations
+#' (yes, 'B' is unfortunate)
+#' @param seed Reproducibility seed
+#' @param graph Plot or not?
+#' @param conformalize Calibrate or not?
+#' @param type_calibration "splitconformal" (default conformal method), "cv1" (do not use),
+#' "loocv" (do not use)
+#' @param gap length of training set for loocv conformal (do not use)
+#' @param agg "mean" or "median" (aggregation method) for
+#' @param vol "constant" or "garch" (type of volatility modeling for calibrated residuals)
+#' @param type_sim "kde", "surrogates", "bootstrap" (type of simulation for calibrated residuals)
+#' @param ...
+#'
+#' @return an object of class 'forecast' with additional information
+#' @export
+#'
+#' @examples
+#'
+#' par(mfrow=c(2, 2))
+#' obj1 <- ahead::fitforecast(AirPassengers)
+#' obj2 <- ahead::fitforecast(AirPassengers, conformalize = TRUE)
+#' plot(AirPassengers)
+#' plot(obj1)
+#' obj2$plot()
+#' obj2$plot("simulations")
+#'
+fitforecast <- function(y,
+                        h = NULL,
+                        pct_train = 0.9,
+                        pct_calibration = 0.5,
+                        method = c("thetaf",
+                                   "arima",
+                                   "ets",
+                                   "tbats",
+                                   "tslm",
+                                   "dynrmf",
+                                   "ridge2f",
+                                   "naive",
+                                   "snaive"),
+                        level = 95,
+                        B = 1000L,
+                        seed = 17223L,
+                        graph = TRUE,
+                        conformalize = FALSE,
+                        type_calibration = c("splitconformal",
+                                             "cv1",
+                                             "loocv"),
+                        gap = 3L,
+                        agg = c("mean", "median"),
+                        vol = c("constant", "garch"),
+                        type_sim = c("kde", "surrogates", "bootstrap"),
+                        ...)
+{
+  method <- match.arg(method)
+  agg <- match.arg(agg)
+  vol <- match.arg(vol)
+  type_sim <- match.arg(type_sim)
+  type_calibration <- match.arg(type_calibration)
+
+  if (is.null(h))
+  {
+    splitted_y_obj <- splitts(y, split_prob = pct_train)
+    y_train <- splitted_y_obj$training
+    y_test <- splitted_y_obj$testing
+    h_test <- length(y_test)
+  } else {
+    y_train <- y
+    h_test <- h
+  }
+
+  tspx <- tsp(y_train)
+  start_preds <- tspx[2] + 1 / tspx[3]
+
+  if (conformalize == TRUE)
+  {
+    splitted_y_train_obj <- splitts(y_train,
+                                    split_prob = pct_calibration)
+    y_train_calibration <- splitted_y_train_obj$training
+    y_val_calibration <- splitted_y_train_obj$testing
+    h_calibration <- length(y_val_calibration)
+  }
+
[…]
