+++ vignettes/fitforecastbench.Rmd
@@ -1,0 +1,30 @@
+---
+title: "Fit and forecast for benchmarking purposes"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Fit and forecast for benchmarking purposes}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+```{r}
+library(ahead)
+library(forecast)
+```
+
+
+```{r fig.width=7.5}
+obj1 <- ahead::fitforecast(AirPassengers)
+obj2 <- ahead::fitforecast(AirPassengers, 
+                           conformalize = TRUE)
+
+par(mfrow=c(1, 2))
+plot(AirPassengers)
+plot(obj1)
+
+par(mfrow=c(1, 2))
+obj2$plot()
+obj2$plot("simulations")
+```
+
