+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: ahead
 Type: Package
 Title: Time Series Forecasting with uncertainty quantification
-Version: 0.11.0
-Date: 2024-03-11
+Version: 0.12.0
+Date: 2024-03-25
 Author: T. Moudiki
 Maintainer: T. Moudiki <thierry.moudiki@gmail.com>
 Description: Univariate and multivariate time series forecasting with uncertainty quantification.
