+++ NAMESPACE
@@ -20,6 +20,7 @@ export("armagarchf")
 export("basicf")
 export("eatf")
 export("dynrmf")
+export("fitforecast")
 export("loessf")
 export("loocvridge2f")
 export("ridge2f")
