+++ NAMESPACE
@@ -16,7 +16,6 @@ importFrom("stats", "as.ts", "complete.cases", "ecdf", "end", "frequency",
            "ts", "tsp", "tsp<-")
 importFrom("utils", "tail")
 importFrom("utils", "install.packages", "installed.packages")
-# interface
 export("armagarchf")
 export("basicf")
 export("eatf")
@@ -27,7 +26,6 @@ export("loocvridge2f")
 export("ridge2f")
 export("ridge")
 export("varf")
-# utils
 export("createtrendseason")
 export("getreturns")
 export("getsimulations")
