### 2024-04-22 #6477 @crvernon
> Thanks @crvernon , happy to help, although the affiliation seems a little off for this author.

Thanks @spholmes!  Yes, I believe independent researchers (those not claiming an institution) should use the affiliation "Independent Researcher" with the country following.

@thierrymoudiki see an example here to follow for correcting this:  https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
