### 2024-05-03 #6477 @spholmes
👋 @thiyangt, would you like to take a look at this submission on time series?
Thanks
Susan
