### 2024-05-15 #6477 @spholmes
Hi @jpiburn : would it be possible for you to review this time series package paper for JOSS? Many Thanks
Susan
