### 2024-06-15 #6477 @spholmes
Since we have started the review process, there is a new link where this is
happening here:
https://github.com/openjournals/joss-reviews/issues/6754

On Sat, Jun 15, 2024 at 8:20 PM Thierry Moudiki ***@***.***>
wrote:

> @spholmes <https://github.com/spholmes> happy to be a reviewer for this
> submission.
>
> @crvernon <https://github.com/crvernon> @spholmes
> <https://github.com/spholmes> what happened to this?
>
