### 2024-06-28 #6477 @spholmes
:wave: @RhysPeploe, would be interested in looking at this submission to JOSS?
Many thanks,
Susan
