### 2024-06-28 #6477 @spholmes
Thanks Chris

A sexta, 28/06/2024, 18:22, Chris Vernon ***@***.***>
escreveu:

> @spholmes <https://github.com/spholmes> inviting another potential
> reviewer for you...
>
> 👋 @shahmoradi <https://github.com/shahmoradi> - Would you be willing to
> review this submission to JOSS? We carry out our checklist-driven reviews
> here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
