### 2024-08-23 #6754 @spholmes
Hi @shahmoradi and @thiyangt
Let me know if you are having problems with your checklists, I am not seeing them here.
Could you please generate your checklists with
`@editorialbot generate my checklist`
and check off the COI and as many items as you have the answers to?
Many thanks
