### 2024-09-05 #6754 @spholmes
Hi @jameshorine, I wonder if I can interest you in helping us review this R package on time series?
Many thanks
Susan
