+++ Makefile
@@ -1,0 +1,134 @@
+.PHONY: build buildsite check clean coverage docs getwd initialize install installcranpkg installgithubpkg installedpkgs load removepkg render setwd start test usegit
+.DEFAULT_GOAL := help
+
+# The directory where R files are stored
+R_DIR = ./R
+
+define BROWSER_PYSCRIPT
+import os, webbrowser, sys
+from urllib.request import pathname2url
+
+# The input is expected to be the full HTML filename
+filename = sys.argv[1]
+filepath = os.path.abspath(os.path.join("./vignettes/", filename))
+webbrowser.open("file://" + pathname2url(filepath))
+endef
+export BROWSER_PYSCRIPT
+
+define PRINT_HELP_PYSCRIPT
+import re, sys
+
+for line in sys.stdin:
+	match = re.match(r'^([a-zA-Z_-]+):.*?## (.*)$$', line)
+	if match:
+		target, help = match.groups()
+		print("%-20s %s" % (target, help))
+endef
+export PRINT_HELP_PYSCRIPT
+
+BROWSER := python3 -c "$$BROWSER_PYSCRIPT"
+
+build: setwd ## build package
+	Rscript -e "devtools::build('.')"
+
+buildsite: setwd ## create a website for the package
+	Rscript -e "pkgdown::build_site('.')"
+
+check: clean setwd ## check package 
+	@read -p "Enter options (e.g: --no-tests --no-examples) or leave empty: " pckgcheckoptions; \
+	if [ -z "$$pckgcheckoptions" ]; then \
+		Rscript -e "try(devtools::check('.'), silent=TRUE)" && exit 0; \
+	fi; \
+	Rscript -e "try(devtools::check('.', args=base::strsplit('$$pckgcheckoptions', ' ')[[1]]), silent=TRUE)";
+
+clean: ## remove all build, and artifacts
+	rm -f .Rhistory
+	rm -f *.RData
+	rm -f *.Rproj
+	rm -rf .Rproj.user
+
+coverage: ## get test coverage
+	Rscript -e "devtools::test_coverage('.')"
+
+create: setwd ## create a new package in current directory
+	Rscript -e "usethis::create_package(path = getwd(), rstudio = FALSE)"
+	rm -f .here
+
+docs: clean setwd ## generate docs		
+	Rscript -e "devtools::document('.')"
+
+getwd: ## get current directory
+	Rscript -e "getwd()"
+
+install: clean setwd ## install current package
+	Rscript -e "try(devtools::install('.'), silent = FALSE)"
+
+installcranpkg: setwd ## install a package
+	@read -p "Enter the name of package to be installed: " pckg; \
+	if [ -z "$$pckg" ]; then \
+		echo "Package name cannot be empty."; \
+		exit 1; \
+	fi; \
+	Rscript -e "utils::install.packages('$$pckg', repos='https://cloud.r-project.org')";
+
+installgithubpkg: setwd ## install a package from GitHub ('repository/pkgname')
+	@read -p "Enter the name of package to be installed ('repository/pkgname'): " pckg; \
+	if [ -z "$$pckg" ]; then \
+		echo "Package name cannot be empty."; \
+		exit 1; \
+	fi; \
+	Rscript -e "devtools::install_github('$$pckg')";
+
+installedpkgs: ## list of installed packages
+	Rscript -e "utils::installed.packages()[,c(10, 16)]"
+
+initialize: setwd ## initialize: install packages devtools, usethis, pkgdown and rmarkdown
[…]
