### 2024-09-22 #6477 @spholmes
> @editorialbot generate my checklist

Hi @shahmoradi , the review is in the thread here: https://github.com/openjournals/joss-reviews/issues/6754
