### 2024-10-01 #6754 @jameshorine
Hi @thierrymoudiki I am working through the checklist this week! As a time-series practitioner, I **really** appreciate the work you put into the package.

Outside of the checklist required by JOSS, there are a few stylistic changes I believe would elevate discourse and greatly improve readability.

Documentation:
1. All figures appear to have the caption "png". Do you mind expanding more for the audience? I try to have as much empathy for the readers as possible.
2. The Vignettes are great!

Readability and Stylistic:
1. Draft Line 22: suggestion to phrase as: "The models implemented in `ahead` include..."
2. Draft line 26: What do you think about indicating what enhancements you made? why is this different and compelling? (I think the choice of L1 regularization is SLICK!!! mention it!) Something like: As in NNAR, `dynrmf` automatically selects p-AR components, and P-seasonal AR components (https://otexts.com/fpp2/seasonal-arima.html).
3. Draft line 29: Excellent!
4. Draft line 41-43: stylistic suggestion: In all of the examples presented here, we utilize the default hyper parameters of...xyz. For more details see ref, ref, ref,.
5. Line 71: stylistic suggestion: "The AirPassengers dataset has been widely tested as a specialized benchmark in forecasting literature because it displays trend, seasonality, and heteroskedasticity.
6. AirPassengers example code runs off page - in your subsequent examples you use a carriage return to improve readability! I recommend to use that style across all of your examples!
7. declaring h <- 20L, while it is normal for time series practitioners to instantly think "forecast horizon", it may not be for others. Suggestion to simply add a comment to improve audience readability!
8. In the last example I would love to see the plot function being called from the ahead package - It took me a little while to figure out that you are masking base plot with `ahead::ahead-plotting-functions`. This has significant helpers that act behind the scenes.
9. To be even more explicit with the internals of the dynrmf function - what do you think about changing the `fit_func` to `model`? That may more clearly signal to users and readers that the beauty of this is exactly the flexibility to use a variety of model architectures!!!
