### 2024-10-12 #6754 @jameshorine
@thierrymoudiki - Looking forward to it! Please feel free to reach out with any questions.

@shahmoradi Do you have any additional feedback you can share?
