+++ NEWS.md
@@ -1,3 +1,6 @@
+# version 0.14.0
+
+- Add `fit_func` and `predict_func` for custom fitting and prediction functions of `ahead::dynrmf` (using `caret` Machine Learning).
 
 # version 0.11.0
 
