+++ man/fit_func.Rd
@@ -1,0 +1,41 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/caretfitpredict.R
+\name{fit_func}
+\alias{fit_func}
+\title{Fit univariate time series using caret ML model (for use with \code{dynrmf})}
+\usage{
+fit_func(
+  x,
+  y,
+  method = "ranger",
+  initial_window = 10L,
+  horizon = 10L,
+  fixed_window = FALSE,
+  tune_length = 5,
+  summary_function = NULL,
+  verbose = TRUE
+)
+}
+\arguments{
+\item{x}{A matrix of predictors}
+
+\item{y}{A vector of responses}
+
+\item{method}{The caret method to use for fitting the model}
+
+\item{initial_window}{The initial window size}
+
+\item{horizon}{The forecast horizon}
+
+\item{fixed_window}{Whether to use a fixed window size}
+
+\item{tune_length}{Length of the tuning grid}
+
+\item{verbose}{Whether to print the model summary}
+}
+\value{
+A model object
+}
+\description{
+Fit univariate time series using caret ML model (for use with \code{dynrmf})
+}
