+++ vignettes/fitforecastbench.Rmd
@@ -18,13 +18,32 @@ library(forecast)
 obj1 <- ahead::fitforecast(AirPassengers)
 obj2 <- ahead::fitforecast(AirPassengers, 
                            conformalize = TRUE)
+obj3 <- ahead::fitforecast(USAccDeaths, method = "dynrmf")
+obj4 <- ahead::fitforecast(USAccDeaths, 
+                           conformalize = TRUE, method = "dynrmf")
+obj5 <- ahead::fitforecast(USAccDeaths, method = "te")
+obj6 <- ahead::fitforecast(USAccDeaths, 
+                           conformalize = TRUE, method = "te")
+
 
 par(mfrow=c(1, 2))
 plot(AirPassengers)
 plot(obj1)
 
 par(mfrow=c(1, 2))
+plot(AirPassengers)
+plot(obj2)
+
+par(mfrow=c(1, 2))
 obj2$plot()
 obj2$plot("simulations")
+
+par(mfrow=c(1, 2))
+obj3$plot()
+obj4$plot()
+
+par(mfrow=c(1, 2))
+obj5$plot()
+obj6$plot()
 ```
 
