+++ R/comb_GLMNET.R
@@ -1,0 +1,98 @@
+#' @title GLMNET Regression Forecast Combination
+#'
+#' @description Computes forecast combination weights using GLMNET Regression (OLS) regression.
+#'
+#' @details
+#' The function integrates the GLMNET Regression forecast combination implementation of the
+#' \emph{ForecastCombinations} package into ForecastComb.
+#'
+#' The results are stored in an object of class 'foreccomb_res', for which separate plot and summary functions are provided.
+#'
+#' @param x An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.
+#'
+#' @return Returns an object of class \code{foreccomb_res} with the following components:
+#' \item{Method}{Returns the best-fit forecast combination method.}
+#' \item{Models}{Returns the individual input models that were used for the forecast combinations.}
+#' \item{Weights}{Returns the combination weights obtained by applying the combination method to the training set.}
+#' \item{Intercept}{Returns the intercept of the linear regression.}
+#' \item{Fitted}{Returns the fitted values of the combination method for the training set.}
+#' \item{Accuracy_Train}{Returns range of summary measures of the forecast accuracy for the training set.}
+#' \item{Forecasts_Test}{Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.}
+#' \item{Accuracy_Test}{Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.}
+#' \item{Input_Data}{Returns the data forwarded to the method.}
+#'
+#' @examples
+#' obs <- rnorm(100)
+#' preds <- matrix(rnorm(1000, 1), 100, 10)
+#' train_o<-obs[1:80]
+#' train_p<-preds[1:80,]
+#' test_o<-obs[81:100]
+#' test_p<-preds[81:100,]
+#'
+#' data<-ForecastComb::foreccomb(train_o, train_p, test_o, test_p)
+#' ahead::comb_GLMNET(data)
+#'
+#' @seealso
+#' \code{\link[ForecastCombinations]{Forecast_comb}},
+#' \code{\link{foreccomb}},
+#' \code{\link{plot.foreccomb_res}},
+#' \code{\link{summary.foreccomb_res}},
+#' \code{\link[forecast]{accuracy}}
+#'
+#' @keywords models
+#'
+#' @import forecast
+#'
+#' @export
+comb_GLMNET <- function(x, custom_error = NULL) {
+    if (!inherits(x, "foreccomb"))
+        stop("Data must be class 'foreccomb'. See ?foreccomb, to bring data in correct format.", call. = FALSE)
+    observed_vector <- x$Actual_Train
+    prediction_matrix <- x$Forecasts_Train
+    modelnames <- x$modelnames
+
+    lin_model <- glmnet::cv.glmnet(x = as.matrix(prediction_matrix), y = as.numeric(observed_vector))    
+    weights <- coef(lin_model, s = "lambda.min")
+    intercept <- weights[1]
+    fitted <- drop(predict(lin_model, prediction_matrix, s = "lambda.min"))
+    accuracy_insample <- forecast::accuracy(observed_vector, as.numeric(fitted))
+    if (!is.null(custom_error)) {
+        accuracy_insample <- cbind(accuracy_insample, custom_error(as.numeric(observed_vector), as.numeric(fitted)))
+        colnames(accuracy_insample)[length(accuracy_insample)] <- "Custom Error"
+    }
+
+    if (is.null(x$Forecasts_Test) && is.null(x$Actual_Test)) {
+        result <- foreccomb_res(method = "GLMNET Regression Regression", modelnames = modelnames, weights = weights, intercept = intercept, fitted = fitted, accuracy_insample = accuracy_insample,
[…]
