+++ R/plot.foreccomb_res.R
@@ -1,0 +1,127 @@
+#' @name plot.foreccomb_res
+#' @aliases plot.foreccomb_res
+#'
+#' @title Plot results from forecast combination model
+#' @description Produces plots for the results of a forecast combination method. Either
+#' an actual vs. fitted plot (\code{which = 1}) or a barplot of the combination weights
+#' (\code{which = 2}).
+#'
+#' @param x An object of class 'foreccomb_res'.
+#' @param which Type of plot: 1 = actual vs. fitted, 2 = combination weights.
+#' @param ... Other arguments passing to \code{\link[graphics]{plot.default}}.
+#'
+#' @return A plot for the foreccomb_res class.
+#'
+#' @examples
+#' obs <- rnorm(100)
+#' preds <- matrix(rnorm(1000, 1), 100, 10)
+#' train_o<-obs[1:80]
+#' train_p<-preds[1:80,]
+#' test_o<-obs[81:100]
+#' test_p<-preds[81:100,]
+#'
+#' data<-foreccomb(train_o, train_p, test_o, test_p)
+#' fit <- comb_EIG1(data)
+#' plot(fit)
+#'
+#' @seealso
+#' \code{\link[ForecastComb]{foreccomb}},
+#' \code{\link[ForecastComb]{summary.foreccomb_res}}
+#'
+#' @author adapted from Christoph E. Weiss and Gernot R. Roetzer (ForecastComb)
+#'
+#' @import ggplot2
+#' @importFrom graphics barplot
+#'
+#' @method plot foreccomb_res
+#' @export
+plot.foreccomb_res <- function(x, which=1,...) {
+    if (!inherits(x, "foreccomb_res"))
+        stop("Data must be class 'foreccomb'. See ?foreccomb, to bring data in correct format.", call. = FALSE)
+    method <- x$Method
+    models <- x$Models
+    rolling<-FALSE
+    weights <- x$Weights
+    if (!is.null(dim(weights))){
+      rolling <- TRUE
+      weights <- try(colMeans(weights), silent = TRUE)
+      if (inherits(weights, "try-error"))
+      {
+        weights <- colMeans(matrix(weights, nrow=1))
+      }
+    }
+    fit <- x$Fitted
+    forec <- x$Forecasts_Test
+    observed_vector <- x$Input_Data$Actual_Train
+    newobs_vector <- x$Input_Data$Actual_Test
+    Index <- NULL  #Hack to satisfy CRAN check.
+
+    if (which == 1){
+    if (is.null(forec) & is.null(newobs_vector)) {
+        cols <- c(ACTUAL = "black", `COMBINED (FIT)` = "#F04546")
+
+        pl <- as.data.frame(matrix(NA, ncol = 3, nrow = length(observed_vector)))
+        colnames(pl) <- c("Index", "Actual", "Combined_Fit")
+        pl[, 1] <- 1:nrow(pl)
+        pl[, 2] <- c(observed_vector)
+        pl[, 3] <- fit
+
+        p <- ggplot(data = pl, aes(x = Index)) + geom_line(aes(y = Actual, colour = "ACTUAL"), size = 0.5) + geom_line(aes(y = Combined_Fit, colour = "COMBINED (FIT)"),
+            size = 0.8) + scale_x_continuous(breaks = round(seq(0, max(pl$Index), by = nrow(pl)/10), 0)) + scale_colour_manual(name = "Series", values = cols) + guides(colour = guide_legend(override.aes = list(size = c(0.5,
+            0.8)))) + xlab("Index") + ylab(paste0(method, "\n Fitted Values/Forescasts")) + ggtitle(paste0(method, " Forecast Combination \n Actual vs. Fitted/Test Set Forecasts")) +
[…]
