+++ R/comb_OLS.R
@@ -22,11 +22,11 @@
 #' can cause major shifts of the coefficient vector (\sQuote{bouncing betas}) -- often causing poor out-of-sample performance.
 #' This issue is addressed by the \code{\link{comb_LAD}} method that is more robust to outliers.
 #'
-#' The results are stored in an object of class 'foreccomb_res', for which separate plot and summary functions are provided.
+#' The results are stored in an object of class 'ForecastComb::foreccomb_res', for which separate plot and summary functions are provided.
 #'
 #' @param x An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.
 #'
-#' @return Returns an object of class \code{foreccomb_res} with the following components:
+#' @return Returns an object of class \code{ForecastComb::foreccomb_res} with the following components:
 #' \item{Method}{Returns the best-fit forecast combination method.}
 #' \item{Models}{Returns the individual input models that were used for the forecast combinations.}
 #' \item{Weights}{Returns the combination weights obtained by applying the combination method to the training set.}
@@ -51,8 +51,8 @@
 #' @seealso
 #' \code{\link[ForecastCombinations]{Forecast_comb}},
 #' \code{\link{foreccomb}},
-#' \code{\link{plot.foreccomb_res}},
-#' \code{\link{summary.foreccomb_res}},
+#' \code{\link{plot.ForecastComb::foreccomb_res}},
+#' \code{\link{summary.ForecastComb::foreccomb_res}},
 #' \code{\link[forecast]{accuracy}}
 #'
 #' @keywords models
@@ -79,7 +79,7 @@ comb_OLS <- function(x, custom_error = NULL) {
     }
 
     if (is.null(x$Forecasts_Test) && is.null(x$Actual_Test)) {
-        result <- foreccomb_res(method = "Ordinary Least Squares Regression", modelnames = modelnames, weights = weights, intercept = intercept, fitted = fitted, accuracy_insample = accuracy_insample,
+        result <- ForecastComb::foreccomb_res(method = "Ordinary Least Squares Regression", modelnames = modelnames, weights = weights, intercept = intercept, fitted = fitted, accuracy_insample = accuracy_insample,
                                 input_data = list(Actual_Train = x$Actual_Train, Forecasts_Train = x$Forecasts_Train), 
                                 predict = predict.comb_OLS)
     }
@@ -88,7 +88,7 @@ comb_OLS <- function(x, custom_error = NULL) {
         newpred_matrix <- x$Forecasts_Test
         pred <- as.vector(lin_model$coef %*% t(cbind(1, newpred_matrix)))
         if (is.null(x$Actual_Test) == TRUE) {
-            result <- foreccomb_res(method = "Ordinary Least Squares Regression", modelnames = modelnames, weights = weights, intercept = intercept, fitted = fitted, accuracy_insample = accuracy_insample,
+            result <- ForecastComb::foreccomb_res(method = "Ordinary Least Squares Regression", modelnames = modelnames, weights = weights, intercept = intercept, fitted = fitted, accuracy_insample = accuracy_insample,
                                     pred = pred, input_data = list(Actual_Train = x$Actual_Train, Forecasts_Train = x$Forecasts_Train, Forecasts_Test = x$Forecasts_Test), 
                                     predict = predict.comb_OLS)
         } else {
@@ -98,13 +98,13 @@ comb_OLS <- function(x, custom_error = NULL) {
                 accuracy_outsample <- cbind(accuracy_outsample, custom_error(as.numeric(newobs_vector), as.numeric(pred)))
                 colnames(accuracy_outsample)[length(accuracy_outsample)] <- "Custom Error"
             }
[…]
