+++ R/zzz.R
@@ -3,4 +3,8 @@
  {
     utils::install.packages("ForecastComb", repos="https://cloud.r-project.org")
  }
+ if (!("caret" %in% rownames(utils::installed.packages()))) 
+ {
+    utils::install.packages("caret", repos="https://cloud.r-project.org")
+ }
 }
