+++ vignettes/comb_OLS.Rmd
@@ -28,9 +28,13 @@ train_obs <- electricity[1:84, "Actual"]
 train_pred <- electricity[1:84, forecasting_methods]
 test_obs <- electricity[85:123, "Actual"]
 test_pred <- electricity[85:123, forecasting_methods]
+```
 
+```{r}
 data <- ForecastComb::foreccomb(train_obs, train_pred, test_obs, test_pred)
+```
 
+```{r}
 start <- proc.time()[3]
 obj <- ahead::comb_OLS(data)
 print(proc.time()[3] - start)
