### 2024-10-27 04745f6
add 'forecast', 'ggplot2' to Depends
