### 2024-11-06 #6754 @thierrymoudiki
@spholmes I just don't want to address the same issue twice. Will answer the requests from @jameshorine. The paper is relatively short though... No formulas. That's why I've been  wondering why it takes such a long time. Thanks. Have a good day
