### 2024-11-06 #6754 @jameshorine
@thierrymoudiki please let me know if I can answer any of the questions about my comments. Happy to hop on a call too.

@spholmes I've asked my professional network for some assistance too.
