+++ R/predict.foreccomb_res.R
@@ -16,18 +16,6 @@
 #' one forecast model, set to \code{TRUE}.
 #' @param ... potential further arguments (require by generic)
 #'
-#' @examples
-#' obs <- rnorm(100)
-#' preds <- matrix(rnorm(1000, 1), 100, 10)
-#' train_o<-obs[1:80]
-#' train_p<-preds[1:80,]
-#' test_o<-obs[81:100]
-#' test_p<-preds[81:100,]
-#'
-#' data<-foreccomb(train_o, train_p)
-#' fit<-comb_BG(data)
-#' predict(fit, test_p)
-#'
 #' @seealso
 #' \code{\link[ForecastComb]{foreccomb}},
 #'
