+++ R/plot.foreccomb_res.R
@@ -12,18 +12,6 @@
 #'
 #' @return A plot for the foreccomb_res class.
 #'
-#' @examples
-#' obs <- rnorm(100)
-#' preds <- matrix(rnorm(1000, 1), 100, 10)
-#' train_o<-obs[1:80]
-#' train_p<-preds[1:80,]
-#' test_o<-obs[81:100]
-#' test_p<-preds[81:100,]
-#'
-#' data<-foreccomb(train_o, train_p, test_o, test_p)
-#' fit <- comb_EIG1(data)
-#' plot(fit)
-#'
 #' @seealso
 #' \code{\link[ForecastComb]{foreccomb}},
 #' \code{\link[ForecastComb]{summary.foreccomb_res}}
