### 2024-11-21 cc367c5
vignette with generic forecasting and conformalization
