+++ R/zzz.R
@@ -1,3 +1,8 @@
 .onLoad <- function(libname, pkgname){
- 
+    if (!requireNamespace("devtools", quietly = TRUE)) {
+      utils::install.packages("devtools")
+    }
+    if (!requireNamespace("misc", quietly = TRUE)) {
+    devtools::install_github("thierrymoudiki/misc")
+    }
 }
