+++ DESCRIPTION
@@ -7,7 +7,7 @@ Author: T. Moudiki
 Maintainer: T. Moudiki <thierry.moudiki@gmail.com>
 Description: Univariate and multivariate time series forecasting with uncertainty quantification.
 License: BSD_3_clause Clear + file LICENSE
-Imports: Rcpp (>= 1.0.6)
+Imports: Rcpp (>= 1.0.6), misc
 LinkingTo: Rcpp
 RoxygenNote: 7.3.0
 Encoding: UTF-8
@@ -28,3 +28,4 @@ Suggests:
     roxygen2
 Roxygen: list(markdown = TRUE)    
 VignetteBuilder: knitr
+Remotes: thierrymoudiki/misc
