### 2024-11-22 #6754 @thierrymoudiki
@spholmes @jameshorine:
- please find an updated version of the paper here: https://github.com/Techtonique/ahead/blob/paper/paper-v2.pdf
- not sure if it's the correct way to post a revision, please tell me
- did you find another reviewer? I don't think the person (who, by the way, is pushing commits everyday) mentioned will ever respond (I don't even see time series forecasting on his profile)

@jameshorine I didn't get the 'corridor' thing, but gave more details about the other real-world applications appearing in the first version.
