+++ DESCRIPTION
@@ -26,6 +26,6 @@ Suggests:
     e1071,
     randomForest,
     roxygen2
-Roxygen: list(markdown = TRUE)    
+Roxygen: list(markdown = TRUE)
 VignetteBuilder: knitr
 Remotes: thierrymoudiki/misc
