### 2024-11-28 #6754 @spholmes
@thierrymoudiki : I did email several people about the paper and still do not have any response either from the person who committed or others. People who write Julia code rarely accept to review R packages, so it would be better if you could suggest authors from JOSS papers of R packages who could effectively review and understand the JOSS process.
In the meantime, could you attend to the reference problems?
