### 2024-11-30 #6754 @thierrymoudiki
@spholmes Not every paper has a DOI (and you know that + this error is new, never appeared before), so I have absolutely no idea on how to fix this issue
