### 2024-12-14 #6477 @thierrymoudiki
@spholmes thanks.

PING

> Hi @nelsonroque, I know you published a paper on your time series feature package a few years back, could you help us review this JOSS submission, we have been having trouble finding a final reviewer for the package as so many people are busy, as you know the JOSS process just involves a checklist and some verifications but is not too onerous. Many Thanks (HAPPY THANKSGIVING) Susan
