### 2025-01-20 7cfa08e
fix genericforecast
