### 2025-01-20 c55d679
fix genericforecast Pt.2
