### 2025-01-22 4acf271
fix genforecast
