+++ R/ml.R
@@ -1,169 +1,201 @@
+#' Conformalized Forecasting using Machine Leaning models
 #'
-#' # y <- AirPassengers;
-#' # obj <- model_fit(y, lags=2, obj = ahead::ridge);
-#' # (res <- model_predict(obj, h=20L));
-#' # plot(res, type = 'l')
+#' @param y A numeric vector or time series of class \code{ts}
+#' @param h Forecasting horizon
+#' @param level Confidence level for prediction intervals
+#' @param lags Number of lags of the input time series considered in the regression
+#' @param fit_func Fitting function (Statistical/ML model). Default is Ridge regression.
+#' @param predict_func Prediction function (Statistical/ML model)
+#' @param type_pi Type of prediction interval
+#' @param B Number of bootstrap replications or number of simulations
+#' @param agg "mean" or "median" (aggregation method)
+#' @param ... additional parameters passed to the fitting function \code{fit_func}
 #'
-#' # y <- Nile;
-#' # obj <- model_fit(y, lags=2, obj = ahead::ridge);
-#' # (res <- model_predict(obj, h=20L));
-#' # plot(res, type = 'l')
-#' #
-#' # y <- USAccDeaths;
-#' # obj <- model_fit(y, lags=2, obj = ahead::ridge);
-#' # (res <- model_predict(obj, h=20L));
-#' # plot(res, type = 'l')
-#' #
-#' # y <- mdeaths;
-#' # obj <- model_fit(y, lags=2, obj = ahead::ridge);
-#' # (res <- model_predict(obj, h=20L));
-#' # plot(res, type = 'l')
-#' #
-#' # y <- austres
-#' # obj <- model_fit(y, lags=2, obj = ahead::ridge);
-#' # (res <- model_predict(obj, h=20L));
-#' # plot(res, type = 'l')
+#' @return
+#' @export
 #'
-#'
-#' #' Forecasting using Machine Leaning models
-#' #'
-#' #' @param y A numeric vector or time series of class \code{ts}
-#' #' @param h Forecasting horizon
-#' #' @param level Confidence level for prediction intervals
-#' #' @param lags Number of lags of the input time series considered in the regression
-#' #' @param fit_func Fitting function (Statistical/ML model). Default is Ridge regression.
-#' #' @param predict_func Prediction function (Statistical/ML model)
-#' #' @param type_pi Type of prediction interval
-#' #' @param ... additional parameters passed to the fitting function \code{fit_func}
-#' #'
-#' #' @return
-#' #' @export
-#' #'
-#' #' @examples
-#' mlf <- function(y, h = 5,
-#'                 level = 95,
-#'                 lags = 1L,
-#'                 fit_func = ahead::ridge,
-#'                 predict_func = predict,
-#'                 fit_params = NULL,
-#'                 type_pi = c("gaussian", "E", "A", "T",
-#'                             "kde", "surogate",
-#'                             "bootstrap",
-#'                             "blockbootsrap",
-#'                             "movingblockbootsrap"),
-#'                 ...)
-#' {
-#'
-#'   type_pi <- match.arg(type_pi)
-#'
-#'   fit_obj <- model_fit(y, lags = lags,
-#'                        obj = fit_func,
-#'                        predict_func = predict_func,
-#'                         ...)
-#'   out <- model_predict(fit_obj, h = h,
-#'                        predict_func = predict_func,
-#'                        type_pi = type_pi,
-#'                        level = level,
-#'                        ...)
-#'   out$x <- y
[…]
