+++ R/ridge2.R
@@ -260,43 +260,107 @@ ridge2f <- function(y,
       margins = margins
     )
   } else { # if (type_pi == "splitconformal") # experimental
-    stop("forthcoming")
-    # y_train_calibration <- splitts(y, split_prob=0.5)
-    # y_train <- y_train_calibration$training
-    # y_calibration <- y_train_calibration$testing
-    #
-    # fit_obj_train <- fit_ridge2_mts(
-    #   y_train,
-    #   lags = lags,
-    #   nb_hidden = nb_hidden,
-    #   nodes_sim = nodes_sim,
-    #   activ = activ,
-    #   a = a,
-    #   lambda_1 = lambda_1,
-    #   lambda_2 = lambda_2,
-    #   dropout = dropout,
-    #   centers = centers,
-    #   type_clustering = type_clustering,
-    #   seed = seed
-    # )
-    #
-    # y_pred_calibration <- ts(
-    #   data = fcast_ridge2_mts(
-    #     fit_obj_train,
-    #     h = nrow(y_calibration),
-    #     type_forecast = type_forecast,
-    #     level = level
-    #   ),
-    #   start = start_preds,
-    #   frequency = freq_x
-    # )
-    #
-    # matrix_y_calibration <- matrix(as.numeric(y_calibration), ncol = 2)
-    # matrix_y_pred_calibration <- matrix(as.numeric(y_pred_calibration), ncol = 2)
-    # abs_residuals <- abs(matrix_y_calibration - matrix_y_pred_calibration)
-    #
-    # quantile_absolute_residuals_conformal <- quantile_scp(abs_residuals = abs_residuals,
-    #                                                       alpha = (1 - level/100))
+    # Split the training data
+    y_train_calibration <- splitts(y, split_prob=0.5)
+    y_train <- y_train_calibration$training 
+    y_calibration <- y_train_calibration$testing
+    h_calibration <- nrow(y_calibration)
+
+    # First fit on training data
+    fit_obj_train <- fit_ridge2_mts(
+      y_train,
+      lags = lags,
+      nb_hidden = nb_hidden,
+      nodes_sim = nodes_sim,
+      activ = activ,
+      a = a,
+      lambda_1 = lambda_1,
+      lambda_2 = lambda_2,
+      dropout = dropout,
+      centers = centers,
+      type_clustering = type_clustering,
+      seed = seed
+    )
+
+    # Get predictions on calibration set
+    y_pred_calibration <- ridge2f(
+      y_train,
+      h = h_calibration,
+      lags = lags,
+      nb_hidden = nb_hidden,
+      nodes_sim = nodes_sim,
+      activ = activ,
+      a = a,
+      lambda_1 = lambda_1,
+      lambda_2 = lambda_2,
+      dropout = dropout,
+      centers = centers,
+      type_clustering = type_clustering,
+      seed = seed
+    )$mean
+
+    # Calculate residuals on calibration set
+    matrix_y_calibration <- matrix(as.numeric(y_calibration), ncol = ncol(y))
+    matrix_y_pred_calibration <- matrix(as.numeric(y_pred_calibration), ncol = ncol(y))
+    abs_residuals <- abs(matrix_y_calibration - matrix_y_pred_calibration)
+    
+    # Get conformal quantile for each series
+    quantile_absolute_residuals_conformal <- apply(abs_residuals, 2, function(x) {
+      quantile_scp(
+        abs_residuals = x,
+        alpha = (1 - level/100)
+      )
+    })
+
+    # Final fit and forecast on full calibration set
+    preds <- ridge2f(
+      y_calibration, 
+      h = h,
+      lags = lags,
[…]
