+++ R/dynrm.R
@@ -102,22 +102,22 @@ dynrmf <- function(y, h = 5,
                    xreg_predict = NULL,
                    ...)
 {
-  if(is_package_available("forecast") == FALSE)
-  {
-    try_forecast <- try(utils::install.packages("forecast",
-                                                repos = c("https://cloud.r-project.org",
-                                                          "https://cran.rstudio.com")),
-    silent = TRUE)
-    if(inherits(try_forecast, "try-error"))
-    {
-      try(utils::install.packages("remotes",
-                                  repos = c("https://cloud.r-project.org",
-                                            "https://cran.rstudio.com")))
-      try(remotes::install_github('robjhyndman/forecast',
-                                  dependencies=TRUE),
-          silent=TRUE)
-    }
-  }
+  # if(is_package_available("forecast") == FALSE)
+  # {
+  #   try_forecast <- try(utils::install.packages("forecast",
+  #                                               repos = c("https://cloud.r-project.org",
+  #                                                         "https://cran.rstudio.com")),
+  #   silent = TRUE)
+  #   if(inherits(try_forecast, "try-error"))
+  #   {
+  #     try(utils::install.packages("remotes",
+  #                                 repos = c("https://cloud.r-project.org",
+  #                                           "https://cran.rstudio.com")))
+  #     try(remotes::install_github('robjhyndman/forecast',
+  #                                 dependencies=TRUE),
+  #         silent=TRUE)
+  #   }
+  # }
 
   stopifnot(length(level) == 1)
 
