+++ R/VAR.R
@@ -50,9 +50,9 @@ varf <- function(y, h = 5,
                               "both", "none"),
                  ...)
 {
-  if(is_package_available("vars") == FALSE)
-    install.packages("vars",
-                     repos = c(CRAN = "https://cloud.r-project.org"))
+  # if(is_package_available("vars") == FALSE)
+  #   install.packages("vars",
+  #                    repos = c(CRAN = "https://cloud.r-project.org"))
 
   if (!is.ts(y))
   {
