+++ DESCRIPTION
@@ -1,15 +1,15 @@
 Package: ahead
 Type: Package
 Title: Time Series Forecasting with uncertainty quantification
-Version: 0.15.0
-Date: 2025-02-11
+Version: 0.16.0
+Date: 2025-02-16
 Author: T. Moudiki
 Maintainer: T. Moudiki <thierry.moudiki@gmail.com>
 Description: Univariate and multivariate time series forecasting with uncertainty quantification.
 License: BSD_3_clause Clear + file LICENSE
-Imports: Rcpp (>= 1.0.6), misc
+Imports: Rcpp (>= 1.0.6), misc, devtools
 LinkingTo: Rcpp
-RoxygenNote: 7.3.0
+RoxygenNote: 7.3.2
 Encoding: UTF-8
 Depends: R (>= 3.5.0), foreach, snow, forecast, ggplot2, tseries
 Suggests: 
