### 2025-03-03 #6754 @spholmes
@crvernon : I have asked the author to fix various technical issues but he seems unable to do so.
Is there someone more technically able in the team that can send them a link that could help here?
The main issues seem to be " author affiliation and references/DOIs.
I can't send this out to a second reviewer unless we can read a current pdf.
Many thanks, susan
