### 2025-03-03 #6754 @crvernon
Hi @spholmes, I'll be happy to assist here.

:wave: @thierrymoudiki - Let's see if we can get the paper compile error fixed up.

First, why do you have the pip install command at the top of the paper after the header?  I believe this can be removed.

Next, the following is the way your header should appear:

```
---
title: 'ahead: Univariate and multivariate time series forecasting with uncertainty quantification including simulation approaches'
tags:
- Python
- R
- Julia
- Time series
- Forecasting
- Machine Learning
- Uncertainty
date: "March 10, 2024"
output:
  pdf_document: default
  word_document: default
  html_document:
    df_print: paged
authors:
- name: Thierry Moudiki
  orcid: "0000-0002-9422-5459"
  affiliation: 1
affiliations:
- name: "Independent Researcher"
  index: 1
bibliography: biblio.bib
---
```

Note that I took the formatting out of the title as well as removed the parenthesis.  I believe the following header should fix the first round of issues.  So please set the paper.md file up this way.  Also, I notice that you have multiple versions of the paper logged in your repo?  I'm not sure why this is the case.  You should just be changing your paper.md file unless I am missing something here.

Once you make these fixes, call me out in this thread and I'll take a look.
