### 2025-03-04 #6754 @thierrymoudiki
@spholmes @crvernon Here's the updated version: https://github.com/Techtonique/ahead/blob/paper/paper.md
