### 2025-03-04 #6754 @spholmes
@thierrymoudiki : Please make the updates in the local thread, the location is in this thread, otherwise the editorial bot does not seem to pick up the changes and is still showing all the references wrong in the older version.
Thanks
