### 2025-03-04 #6754 @crvernon
:wave: @spholmes - The branch is now correct I believe.  So @thierrymoudiki if you keep updating the paper branch all should be fine.

Next step to cleaning the paper up is to remove any references from the bib file that are not directly used in the paper.  This will take care of most of that nasty report from your receiving about your references.

Also, when I look at the paper proof generated above, it is way to long.  @thierrymoudiki you will want to have to word count of the paper under 1000 words.  There are also way to many figures and too much code in your paper.  Shift any code and figures that you can to online docs and just call a link to the docs out in the text.
