### 2025-03-05 #6754 @thierrymoudiki
Thanks @crvernon.
@crvernon @spholmes If this is way too long, I don't know what will be left. Maybe, somehow, "long"; there are 1144 words instead of 1000. But "way too long"?
Also, seems like we're missing a review here.
