+++ man/rmultivariate.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sampling.R
+\name{rmultivariate}
+\alias{rmultivariate}
+\title{Simulate multivariate data}
+\usage{
+rmultivariate(
+  data,
+  method = c("bootstrap", "block-bootstrap"),
+  n = 100L,
+  block_size = 5
+)
+}
+\arguments{
+\item{data}{A numeric vector or matrix.}
+
+\item{method}{The method to use for sampling.}
+
+\item{n}{The number of samples to draw.}
+
+\item{block_size}{The size of the blocks to use for the block bootstrap.}
+
+\item{...}{Additional arguments to pass to the density function.}
+}
+\description{
+Simulate multivariate data
+}
