### 2025-03-05 6052011
vignette for conformal ridge2
