### 2025-03-06 103c281
rm class forecast from mts
