+++ R/ridge2.R
@@ -199,12 +199,9 @@ ridge2f <- function(y,
   if (!is.null(xreg))
   {
     if (is.null(ncol(xreg)))
-      xreg <- as.matrix(xreg)[seq_len(nrow(y)),]
-
-    #stopifnot(identical(nrow(xreg), nrow(y)))
+      xreg <- matrix(xreg, ncol=1)
 
     use_xreg <- TRUE
-
     n_xreg <- dim(xreg)[2]
     colnames_xreg <- colnames(xreg)
     if (is.null(colnames_xreg))
@@ -323,6 +320,36 @@ ridge2f <- function(y,
        } else {
         quantile_absolute_residuals_conformal <- quantile(abs(calibrated_residuals), probs = level/100)
        }
+       
+      # Create output with proper time series attributes
+      out <- list(
+        mean = ts(preds, start = start_preds, frequency = freq_x),
+        lower = ts(preds - quantile_absolute_residuals_conformal, 
+                  start = start_preds, frequency = freq_x),
+        upper = ts(preds + quantile_absolute_residuals_conformal, 
+                  start = start_preds, frequency = freq_x),
+        sims = NULL,
+        x = y,
+        level = level,
+        method = "ridge2",
+        residuals = ts(calibrated_residuals, start = start_x, frequency = freq_x),
+        coefficients = fit_obj_train$coef,
+        loocv = NULL,
+        weighted_loocv = NULL, 
+        loocv_per_series = NULL
+      )
+
+      # Handle univariate case
+      if (dimensionality == "univariate") {
+        for (i in 1:length(out)) {
+          try_delete_trend <- try(delete_columns(out[[i]], "trend_univariate"), silent = TRUE)
+          if (!inherits(try_delete_trend, "try-error") && !is.null(out[[i]])) {
+            out[[i]] <- try_delete_trend
+          }
+        }
+        return(structure(out, class = "forecast"))
+      }
+      return(structure(out, class = "mtsforecast"))
     }
 
     if (type_pi %in% c("conformal-bootstrap", "conformal-block-bootstrap")) {
@@ -401,7 +428,7 @@ ridge2f <- function(y,
         x = y,
         level = level,
         method = "ridge2",
-        residuals = ts(calibrated_residuals, start = start_x),
+        residuals = ts(calibrated_residuals, start = start_x, frequency = freq_x),
         coefficients = fit_obj_train$coef,
         loocv = NULL,
         weighted_loocv = NULL,
@@ -419,67 +446,7 @@ ridge2f <- function(y,
         return(structure(out, class = "forecast"))
       }
       return(structure(out, class = "mtsforecast"))
-    } else { # type_pi == "conformal-split"
-      out <- list(
-        mean = ts(preds, start = start_preds, frequency = freq_x),
-        lower = ts(preds - quantile_absolute_residuals_conformal, start = start_preds, frequency = freq_x),
-        upper = ts(preds + quantile_absolute_residuals_conformal, start = start_preds, frequency = freq_x),
-        sims = NULL,
-        x = y,
-        level = level,
-        method = "ridge2",
-        residuals = calibrated_residuals,
-        coefficients = fit_obj_train$coef,
-        loocv = NULL,
-        weighted_loocv = NULL, 
-        loocv_per_series = NULL
-      )
-      # Handle univariate case
-      if (dimensionality == "univariate") {
-        for (i in 1:length(out)) {
-          try_delete_trend <- try(delete_columns(out[[i]], "trend_univariate"), silent = TRUE)
-          if (!inherits(try_delete_trend, "try-error") && !is.null(out[[i]])) {
-            out[[i]] <- try_delete_trend
[…]
