### 2025-03-07 99b5204
add fanchart and spaghetti plots
