+++ vignettes/ahead-ridge2-conformal.Rmd
@@ -16,15 +16,15 @@ library(ahead)
 ## AirPassengers
 
 ```{r, fig.width=5}
-plot(ahead::ridge2f(AirPassengers, h=20L, lags=20L, type_pi="conformal-split"))
+plot(ahead::ridge2f(AirPassengers, h=30L, lags=20L, type_pi="conformal-split"))
 ```
 
 ```{r, fig.width=5}
-plot(ahead::ridge2f(AirPassengers, h=20L, lags=20L, type_pi="conformal-block-bootstrap"))
+plot(ahead::ridge2f(AirPassengers, h=30L, lags=20L, type_pi="conformal-block-bootstrap"))
 ```
 
 ```{r, fig.width=5}
-plot(ahead::ridge2f(AirPassengers, h=20L, lags=20L, type_pi="conformal-bootstrap"))
+plot(ahead::ridge2f(AirPassengers, h=30L, lags=20L, type_pi="conformal-bootstrap"))
 ```
 
 ## USAccDeaths
@@ -74,4 +74,9 @@ obj <- ahead::ridge2f(fpp::insurance,
 
 plot(obj, "Quotes")
 plot(obj, "TV.advert")
+```
+
+```{r, fig.width=7.2}
+plot(obj, "Quotes", type = "dist")
+plot(obj, "TV.advert", type = "sims")
 ```
