### 2025-03-07 #6754 @spholmes
Hi @bsiepe,
I am looking for a reviewer to look over this submission about the R package "ahead" that provides time series forecasting with uncertainty. Our procedure is very streamlined and checklist based and the papers are quite short,
would you mind taking a look...
Many thanks
Susan
