### 2025-03-08 #6754 @bsiepe
Hi @spholmes,
Thanks for inviting me to review this submission. I'd be happy to review this paper, but I am currently traveling and would need ~2 weeks to finish the review. If that is okay with you, I can take it on.
