### 2025-03-12 e209a81
fix trend in adjusted trend
