+++ R/glmtheta.R
@@ -6,12 +6,16 @@
 #' @param fit_func The function to use for fitting the GLM
 #' @param fan Logical flag for fan plot
 #' @param x The time series data
+#' @param attention Logical flag for using attention mechanism
 #' @param ... Additional arguments to pass to the fit_func
 #' @return A forecast object
 #' @export
 glmthetaf <- function (y, h = ifelse(frequency(y) > 1, 2 * frequency(y), 10), 
-                       level = c(80, 95), fit_func = lsfit, fan = FALSE, x = y, ...) 
+                       level = c(80, 95), fit_func = lsfit, fan = FALSE, x = y, 
+                       attention = FALSE, method=c("base", "adj"),
+                       ...) 
 {
+  method <- match.arg(method) 
   if (fan) {
     level <- seq(51, 99, by = 3)
   }
@@ -45,34 +49,104 @@ glmthetaf <- function (y, h = ifelse(frequency(y) > 1, 2 * frequency(y), 10),
     }
   }
   
+  fcast <- ses(x, h = h) 
+  alpha <- pmax(1e-10, fcast$model$par["alpha"])
+  
   # Try formula interface first, then matrix interface
   time_idx <- 0:(n - 1)
   df <- data.frame(y = x, t = time_idx)
-  tmp2 <- try(fit_func(y ~ t, data = df, ...), silent = TRUE)
-  if (inherits(tmp2, "try-error")) {
-    tmp2 <- try(fit_func(x = matrix(time_idx, ncol=1), y = as.numeric(x), ...), silent = TRUE)
+
+  if (attention == FALSE) {
+    tmp2 <- try(fit_func(y ~ t, data = df, ...), silent = TRUE)
     if (inherits(tmp2, "try-error")) {
-      stop("Unable to fit linear trend")
-    }
-  }
-  
-  # Extract coefficient with multiple fallback methods
-  slope <- try(coef(tmp2)[2], silent = TRUE)
-  if (inherits(slope, "try-error")) {
-    slope <- try(tmp2$coefficients[2], silent = TRUE)
-    if (inherits(slope, "try-error")) {
-      slope <- try(tmp2$coef[2], silent = TRUE)
-      if (inherits(slope, "try-error")) {
-        stop("Unable to extract slope coefficient")
+      # For matrix interface, include intercept term for methods like glmnet
+      X <- matrix(time_idx, nrow=1)
+      tmp2 <- try(fit_func(x = X, y = as.numeric(x), ...), silent = TRUE)
+      if (inherits(tmp2, "try-error")) {
+        stop("Unable to fit linear trend")
       }
     }
+    # Extract coefficient using generic method first, then fallback
+    slope <- try(coef(tmp2)[2], silent = TRUE)
+    if (inherits(slope, "try-error")) {
+      slope <- try(tmp2$coefficients[2], silent = TRUE)
+      if (inherits(slope, "try-error")) {
+        slope <- try(tmp2$coef[2], silent = TRUE)
+        if (inherits(slope, "try-error")) {
+          stop("Unable to extract slope coefficient")
+        }
+      }
+    }
+    tmp2 <- slope/2  # Divide by 2 as per theta method
+  } else { # attention is TRUE
+      if (method == "base")
+      {
+        df <- cbind.data.frame(df, ctx = ahead::computeattention(x)$context_vectors)
+        tmp2 <- try(fit_func(y ~ ., data = df, ...), silent = FALSE)
+        if (inherits(tmp2, "try-error")) {
+          X <- cbind.data.frame(time_idx, df$ctx)
+          tmp2 <- try(fit_func(x = as.matrix(X), y = as.numeric(x), ...), silent = FALSE)
+          if (inherits(tmp2, "try-error")) {
[…]
