+++ vignettes/glmthetaf-attention.Rmd
@@ -1,0 +1,156 @@
+---
+title: "Generalized Linear Model Theta Forecast with attention"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Generalized Linear Model Theta Forecast with attention}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+# USAccDeaths (method='base')
+
+```{r fig.width=7.5}
+library(forecast)
+library(ahead)
+
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::glm.nb, attention = TRUE)))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::glm, attention = TRUE)))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::rlm, attention = TRUE)))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::lqs, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::lsfit, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::lm, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=gam::gam, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=quantreg::rq, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=glmnet::glmnet, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=glmnet::cv.glmnet, attention = TRUE)))
+plot(obj)
+```
+
+# AirPassengers (method='base')
+
+```{r fig.width=5.5}
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=MASS::glm.nb, attention = TRUE)))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=stats::glm, attention = TRUE)))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=MASS::rlm, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=5.5}
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=stats::lm, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=5.5}
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=MASS::lqs, attention = TRUE)))
+plot(obj)
+```
+
+```{r fig.width=5.5}
[…]
