+++ vignettes/glmthetaf-attention-conformal.Rmd
@@ -1,0 +1,156 @@
+---
+title: "Generalized Linear Model Theta Forecast with attention - conformal uncertainty"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Generalized Linear Model Theta Forecast with attention - conformal uncertainty}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+# USAccDeaths (method='base')
+
+```{r fig.width=7.5}
+library(forecast)
+library(ahead)
+
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::glm.nb, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::glm, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::rlm, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::lqs, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::lsfit, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::lm, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=gam::gam, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=quantreg::rq, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=glmnet::glmnet, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+```
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=glmnet::cv.glmnet, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+```
+
+# AirPassengers (method='base')
+
+```{r fig.width=5.5}
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=MASS::glm.nb, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=stats::glm, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=MASS::rlm, attention = TRUE, type_pi = "conformal-split")))
+plot(obj)
+```
+
+```{r fig.width=5.5}
[…]
