+++ vignettes/glmthetaf-attention.Rmd
@@ -13,79 +13,79 @@ vignette: >
 library(forecast)
 library(ahead)
 
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::glm.nb, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::glm.nb, attention = TRUE, method = 'base')))
 plot(obj)
 
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::glm, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::glm, attention = TRUE, method = 'base')))
 plot(obj)
 
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::rlm, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::rlm, attention = TRUE, method = 'base')))
 plot(obj)
 
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::lqs, attention = TRUE)))
-plot(obj)
-```
-
-```{r fig.width=7.5}
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::lsfit, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::lqs, attention = TRUE, method = 'base')))
 plot(obj)
 ```
 
 ```{r fig.width=7.5}
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::lm, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::lsfit, attention = TRUE, method = 'base')))
 plot(obj)
 ```
 
 ```{r fig.width=7.5}
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=gam::gam, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=stats::lm, attention = TRUE, method = 'base')))
 plot(obj)
 ```
 
 ```{r fig.width=7.5}
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=quantreg::rq, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=gam::gam, attention = TRUE, method = 'base')))
 plot(obj)
 ```
 
 ```{r fig.width=7.5}
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=glmnet::glmnet, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=quantreg::rq, attention = TRUE, method = 'base')))
 plot(obj)
 ```
 
 ```{r fig.width=7.5}
-(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=glmnet::cv.glmnet, attention = TRUE)))
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=glmnet::glmnet, attention = TRUE, method = 'base')))
+plot(obj)
+```
+
[…]

[…]
