### 2025-03-13 911464f
fix attention==FALSE Pt.2
