### 2025-03-18 #6754 @spholmes
Thanks so much @bsiepe , I hope you are having safe travels. I am now assigning you as a reviewer and you can generate a checklist easily by typing "@editorialbot generate my checklist", the first thing to check is if you have any COI, otherwise all the rest is pretty straightforward and let me know if you have any problems. Susan
