### 2025-03-20 Techtonique/ahead#24 @bsiepe
Typos in manuscript
During my [JOSS review](https://github.com/openjournals/joss-reviews/issues/6754), I found some typos in the manuscript :

1. l.33 "neural networks" -> should be neural network
2. l. 36: "independen" -> should be independent
3. l.37: missing space
4. l.46: "are used" is redundant
5. l.82: model's hyperparameters: without 's'
