### 2025-03-20 #6754 @bsiepe
First, @thierrymoudiki, I appreciate the work you have put into writing a concise paper and developing your package!

Here are a couple of points of potential improvement and/or clarification:

## General checks
- **License:** Could you clarify whether the "clear" license you use is "The 3-Clause BSD License" as approved by the [OSI](https://opensource.org/license/bsd-3-clause)? Also, the hyperlink to your license on your docs landing page [is down](https://docs.techtonique.net/ahead/LICENSE).

## Documentation:
- **Community guidelines:** The [contribution guidelines](https://github.com/Techtonique/ahead/blob/main/CONTRIBUTING.md) say something about general contributions, but nothing about how to report issues/seek support
- **Functionality documentation:** Documentation for functions is available, but generally sparse. For example, looking at [conformalization](https://docs.techtonique.net/ahead/reference/conformalize.html), there is little explanation of what this approach does. The related vignettes then also mainly show how these results can be plotted. I'm not sure what the JOSS standard are here, but some longer descriptions of the main functions of the package in the docs would be helpful for users.

## Software Paper:
- **Summary:** The paper title presents the `ahead` package, but then mainly focuses on two implemented models. Does the package itself have a broader purpose? If so, what exactly is it? Extending machine learning forecasting techniques with uncertainty quantification? I think briefly clarifying this in the paper and/or in the package docs would be useful.
- **State of the field:** Related to my point about sparse documentation, what exactly is the contribution for `dynrmf`? The paper claims that it was inspired by NNAR, but is more generic. How exactly is it more generic? The package documentation itself also does not explain this point. While `ridge2f` is explained in a paper, the package documentation merely explains that it is a "Random Vector functional link network model with 2 regularization parameters" without any additional explanations (see [here](https://docs.techtonique.net/ahead/reference/ridge2f.html)), which also seems very brief for the main functionality of a package. I think users would profit from more documentation here.
- **Quality of writing:** There are a couple of typos, which I address in this issue https://github.com/Techtonique/ahead/issues/24.

**Misc:**
- l.13: It's not clear what the "M" in MTS stands for - multivariate?
- The figures are all .png files, and some of them have a pretty bad resolution (e.g. the [last figure](https://github.com/Techtonique/ahead/blob/paper/paper_files/paper_30_1.png)). Why not use vector graphics?

**References:**
- There are multiple issues with citations in the manuscript, where parentheses are closed twice (e.g. l.38)
- hyndman2013forecasting has redundant white space in the URL

## Tests
I could run all without fail, except for the [`ridge2f`](https://github.com/Techtonique/ahead/blob/main/tests/testthat/test_ridge2f.R) function, where I get the following error:
```r
res41 <- ahead::ridge2f(x, type_pi = "rvinecopula", B = 5L,
                        margins = "empirical")
```

> Error in params_distro$distro : $ operator is invalid for atomic vectors
