### 2025-03-20 #6754 @thierrymoudiki
@bsiepe Thanks for the review.

> First, [@thierrymoudiki](https://github.com/thierrymoudiki), I appreciate the work you have put into writing a concise paper and developing your package!
>
> Here are a couple of points of potential improvement and/or clarification:
>
> ## General checks
> * **License:** Could you clarify whether the "clear" license you use is "The 3-Clause BSD License" as approved by the [OSI](https://opensource.org/license/bsd-3-clause)? Also, the hyperlink to your license on your docs landing page [is down](https://docs.techtonique.net/ahead/LICENSE).
>

Here is a clarification: https://opensource.stackexchange.com/questions/7559/patent-rights-bsd-3-clause-clear-vs-bsd-3-clause

> ## Documentation:
> * **Community guidelines:** The [contribution guidelines](https://github.com/Techtonique/ahead/blob/main/CONTRIBUTING.md) say something about general contributions, but nothing about how to report issues/seek support

About issues/seek support: IMHO, that's more **general GitHub knowledge** than specific to any package. Since package is hosted here... Github rules apply here.

> * **Functionality documentation:** Documentation for functions is available, but generally sparse. For example, looking at [conformalization](https://docs.techtonique.net/ahead/reference/conformalize.html), there is little explanation of what this approach does. The related vignettes then also mainly show how these results can be plotted. I'm not sure what the JOSS standard are here, but some longer descriptions of the main functions of the package in the docs would be helpful for users.

Can improve the docs (but it will never be perfect; that's always a work in progress). E.g `conformalize` is very recent. **The paper's been under review for more than 1 year** and initially didn't contain this function. **Not sure how far ahead's the target here**.

>
> ## Software Paper:
> * **Summary:** The paper title presents the `ahead` package, but then mainly focuses on two implemented models. Does the package itself have a broader purpose? If so, what exactly is it? Extending machine learning forecasting techniques with uncertainty quantification? I think briefly clarifying this in the paper and/or in the package docs would be useful.

The paper has to be concise. Can't reasonably describe everything the package does; which is: **univariate and multivariate time series forecasting with uncertainty quantification**. "Extending machine learning forecasting techniques with uncertainty quantification" is a subset of this description.

> * **State of the field:** Related to my point about sparse documentation, what exactly is the contribution for `dynrmf`? The paper claims that it was inspired by NNAR, but is more generic. How exactly is it more generic?

As **clearly** demonstrated in the examples, and differently enough from NNAR (which uses one-layered artificial neural networks on automatically determined lags; autoregressive and seasonal),  `dynrmf` can use **any** (i.e _generic_) Machine Learning model you could think of, on automatically determined lags; autoregressive and seasonal. How cool of a contribution is that?

> * The package documentation itself also does not explain this point. While `ridge2f` is explained in a paper, the package documentation merely explains that it is a "Random Vector functional link network model with 2 regularization parameters" without any additional explanations (see [here](https://docs.techtonique.net/ahead/reference/ridge2f.html)), which also seems very brief for the main functionality of a package. I think users would profit from more documentation here.
> * **Quality of writing:** There are a couple of typos, which I address in this issue [Typos in manuscript Techtonique/ahead#24](https://github.com/Techtonique/ahead/issues/24).

Will check and fix when necessary.

>
> **Misc:**
>
> * l.13: It's not clear what the "M" in MTS stands for - multivariate?

Indeed. Will mention 'M'  = 'Multivariate'

> * The figures are all .png files, and some of them have a pretty bad resolution (e.g. the [last figure](https://github.com/Techtonique/ahead/blob/paper/paper_files/paper_30_1.png)). Why not use vector graphics?

Bad = ?. Do you have a metric? What is 'vector graphics'? Can't do better than R, unless you suggest a better way.

> **References:**
>
> * There are multiple issues with citations in the manuscript, where parentheses are closed twice (e.g. l.38)
> * hyndman2013forecasting has redundant white space in the URL

Will check but that's kind of automatic. Redundant white space = ?

> ## Tests
> I could run all without fail, except for the [`ridge2f`](https://github.com/Techtonique/ahead/blob/main/tests/testthat/test_ridge2f.R) function, where I get the following error:
>
> res41 <- ahead::ridge2f(x, type_pi = "rvinecopula", B = 5L,
>                         margins = "empirical")
> > Error in params_distro$distro : $ operator is invalid for atomic vectors

Will check. Have updated the tests in new release. The thing is: the package is evolving, and things change while it's being reviewed.
