### 2025-03-24 #6754 @bsiepe
> [@bsiepe](https://github.com/bsiepe) Thanks for the review.
>
> > First, [@thierrymoudiki](https://github.com/thierrymoudiki), I appreciate the work you have put into writing a concise paper and developing your package!
> > Here are a couple of points of potential improvement and/or clarification:
> > ## General checks
> >
> > * **License:** Could you clarify whether the "clear" license you use is "The 3-Clause BSD License" as approved by the [OSI](https://opensource.org/license/bsd-3-clause)? Also, the hyperlink to your license on your docs landing page [is down](https://docs.techtonique.net/ahead/LICENSE).
>
> Here is a clarification: https://opensource.stackexchange.com/questions/7559/patent-rights-bsd-3-clause-clear-vs-bsd-3-clause

Thanks. This license does not seem to be approved by the OSI (as required by JOSS), but seems very close to approved licenses. I think it's best for the journal editors to decide on this matter.

> > ## Documentation:
> >
> > * **Community guidelines:** The [contribution guidelines](https://github.com/Techtonique/ahead/blob/main/CONTRIBUTING.md) say something about general contributions, but nothing about how to report issues/seek support
>
> About issues/seek support: IMHO, that's more **general GitHub knowledge** than specific to any package. Since package is hosted here... Github rules apply here.

Adding information about how to report issues or seek support is required by JOSS, as you can see in the checklist above.
>
> > * **Functionality documentation:** Documentation for functions is available, but generally sparse. For example, looking at [conformalization](https://docs.techtonique.net/ahead/reference/conformalize.html), there is little explanation of what this approach does. The related vignettes then also mainly show how these results can be plotted. I'm not sure what the JOSS standard are here, but some longer descriptions of the main functions of the package in the docs would be helpful for users.
>
> Can improve the docs (but it will never be perfect; that's always a work in progress). E.g `conformalize` is very recent. **The paper's been under review for more than 1 year** and initially didn't contain this function. **Not sure how far ahead's the target here**.

I understand your frustration with the length of the review process, and I understand that docs are always a work in progress. Still, my comment applies to multiple functions in the package, including those discussed below.

>
> > ## Software Paper:
> >
> > * **Summary:** The paper title presents the `ahead` package, but then mainly focuses on two implemented models. Does the package itself have a broader purpose? If so, what exactly is it? Extending machine learning forecasting techniques with uncertainty quantification? I think briefly clarifying this in the paper and/or in the package docs would be useful.
>
> The paper has to be concise. Can't reasonably describe everything the package does; which is: **univariate and multivariate time series forecasting with uncertainty quantification**. "Extending machine learning forecasting techniques with uncertainty quantification" is a subset of this description.
>
> > * **State of the field:** Related to my point about sparse documentation, what exactly is the contribution for `dynrmf`? The paper claims that it was inspired by NNAR, but is more generic. How exactly is it more generic?
>
> As **clearly** demonstrated in the examples, and differently enough from NNAR (which uses one-layered artificial neural networks on automatically determined lags; autoregressive and seasonal), `dynrmf` can use **any** (i.e _generic_) Machine Learning model you could think of, on automatically determined lags; autoregressive and seasonal. How cool of a contribution is that?

FWIW: I appreciate your work and I am trying to give constructive and helpful feedback such that your package complies with requirements by JOSS and is accessible for users. You can certainly disagree with me, but replying that your contribution is cool and everything is "**clearly**" explained does not necessarily strike me as a highly constructive response.

I'm also not aiming to diminish your contribution, but it would be easier for others to evaluate its coolness if there was more documentation. Why not explicate what exactly you contributed instead of merely writing that it is based on NNAR?
The examples that you mention - unless I am missing any, please point me to them if so - showcase code with different fitting functions, without any further explanations. That is nice, but still does not really explain the implementation itself.

For example:
- Did you just replace the fitting function in NNAR? It does not seem like it, because you offer additional options such as lag selection via AR models.
- In NNAR, it seems to me like uncertainty intervals are based on simulations, whereas you rely on different techniques such as ETS/ARIMA/.. to compute uncertainty intervals. Is that right? Is it documented somewhere?
- Are there any minimum requirements on which models can be used in the function? Could you implement, for example, use a multivariate method or an LSTM? What happens if the fitting and the prediction function differ?
- How does the method deal with seasonality?

As the user of a modeling package, it would be very helpful to have such knowledge readily available in a brief summary. For example, check the documentation of `forecast::nnetar` [here](https://pkg.robjhyndman.com/forecast/reference/nnetar.html)  which provides longer and more detailed explanations.

> > * The package documentation itself also does not explain this point. While `ridge2f` is explained in a paper, the package documentation merely explains that it is a "Random Vector functional link network model with 2 regularization parameters" without any additional explanations (see [here](https://docs.techtonique.net/ahead/reference/ridge2f.html)), which also seems very brief for the main functionality of a package. I think users would profit from more documentation here.
> > * **Quality of writing:** There are a couple of typos, which I address in this issue [Typos in manuscript Techtonique/ahead#24](https://github.com/Techtonique/ahead/issues/24).
>
> Will check and fix when necessary.
>
> > **Misc:**
> >
> > * l.13: It's not clear what the "M" in MTS stands for - multivariate?
>
> Indeed. Will mention 'M' = 'Multivariate
>
> > * The figures are all .png files, and some of them have a pretty bad resolution (e.g. the [last figure](https://github.com/Techtonique/ahead/blob/paper/paper_files/paper_30_1.png)). Why not use vector graphics?
>
> Bad = ?. Do you have a metric? What is 'vector graphics'? Can't do better than R, unless you suggest a better way.

The graphics are somewhat blurry. See [here](https://en.wikipedia.org/wiki/Vector_graphics) for an explanation of vector graphics. In R, you could use .svg or .pdf. Alternatively, you could still increase the resolution of your .png.

> > **References:**
> >
> > * There are multiple issues with citations in the manuscript, where parentheses are closed twice (e.g. l.38)
> > * hyndman2013forecasting has redundant white space in the URL
>
> Will check but that's kind of automatic. Redundant white space = ?

Redundant white space refers to unnecessary spaces in the URL [here](https://github.com/Techtonique/ahead/blob/paper/biblio.bib#L355).
You can just omit the irrelevant closing parentheses where necessary.
>
> > ## Tests
> > I could run all without fail, except for the [`ridge2f`](https://github.com/Techtonique/ahead/blob/main/tests/testthat/test_ridge2f.R) function, where I get the following error:
> > res41 <- ahead::ridge2f(x, type_pi = "rvinecopula", B = 5L,
> > margins = "empirical")
> > > Error in params_distro$distro : $ operator is invalid for atomic vectors
>
> Will check. Have updated the tests in new release. The thing is: the package is evolving, and things change while it's being reviewed. Plus, not all the function features are failing, but this specific test.

Checking the tests is also something required in the JOSS checklist. I understand that the package is in continuous development and I'm aware that not all function features are failing. Still, for a stable download version, it would be great if no tests failed.
