### 2025-03-25 #6754 @spholmes
Thanks @bsiepe for the help in reviewing here, these are very valid comments.
@thierrymoudiki : At JOSS we need to have frozen version of the paper and code that runs all tests and that is what is archived with a DOI.
@crvernon: I need a little help here on some of the technical issues about the JOSS license, is this an issue?
