### 2025-03-25 #6754 @thierrymoudiki
@spholmes Some comments are valid (those about license, tests), others are overzealous (docs, methods). All this energy for a 5-page piece sounds a bit exaggerated (maybe not for the last reviewer who just joined). Some papers in JOSS only have 2 pages. **This isn't a niche forecasting journal**, and it should be reminded. Also, demands ("more", "somewhat") should be extremely precise to avoid hundreds of back-and-forth interactions about "beauty" standards.
