### 2025-03-27 #6754 @thierrymoudiki
@spholmes @crvernon Just updated the paper with the new version (was `0.11.0`, is now... `0.18.0`), along with passing tests (run `make test` in root folder), fixed typos, and ways to reproduce the figures: https://github.com/Techtonique/ahead/blob/paper/vignettes/paper.Rmd (this is my, and everyone's first time hearing that resolution of R figures isn't _good_ enough).

That's my **peak artistic performance**. If there are still substantial changes to be made at this point, for this 5-page-1000-word piece, and after a lengthy 1 year marathon: **just reject**.
