### 2025-04-07 68de18e
fix vignettes for attention=FALSE
