### 2025-04-08 #6754 @bsiepe
The paper does not seem to render due to an issue with the bibliography at the moment, which means that I cannot assess if all citations are correctly included.  It might have to do with moving the `biblio.bib` file to the `/vignettes/` folder?

# Summary of Changes

These are the main changes that are apparent to me when checking the file history:
- I see that the functionality of the package and the two main methods that are used in the paper was clarified a bit more.
- I cannot see the plots in the rendered PDF, but it seems like they have higher resolution based on the .md file. (This issue actually comes up quite frequently in reviews at other peer-reviewed journals).

I have updated my checklist above accordingly.

# What seems not to have changed

Some previous comments of mine seem to have not been adressed (but please feel free to correct me, it's a bit difficult to see all changes across branches):
- The license of the package still may be incompatible with the requirements of JOSS.
- There still is no contributors guide in a format as required by JOSS.
- When installing the package from the main GitHub branch ([commit version](https://github.com/Techtonique/ahead/commit/083c32a1b046b940d74eab165d16ca15dda3da52)) and running all tests, one test still fails:

```R
Error (test_ridge2f.R:20:1): (code run outside of `test_that()`)
Error in `params_distro$distro`: $ operator is invalid for atomic vectors
Backtrace:
    ▆
 1. └─ahead::ridge2f(x, type_pi = "rvinecopula", B = 5L) at test_ridge2f.R:20:1
 2.   └─base::lapply(...) at ahead/R/ridge2.R:812:7
 3.     └─ahead (local) FUN(X[[i]], ...)
 4.       ├─stats::ts(...) at ahead/R/ridge2.R:812:7
 5.       │ └─base::is.data.frame(data)
 6.       └─ahead:::fcast_ridge2_mts(...) at ahead/R/ridge2.R:812:7
 7.         └─ahead:::simulate_rvine(...) at ahead/R/ridge2.R:1354:7

```
Did you run the tests on the main branch or on the paper branch? Maybe there is a misunderstanding here.

- Many of my previous comments regarding documentation concerned the package documentation of functions such as `dynrmf` (also part of the review process at JOSS), not only the paper itself. This documentation seems not to have changed at all from what I see? While I still think this documentation could be improved, as I've specifically outlined above, I don't think additional back and forths are productive. I would thus like to delegate this decision to the editors/other reviewer.
