### 2025-04-09 faf7327
fix vignettes -- glmthetaf
