### 2025-04-11 #6754 @crvernon
@spholmes,  @bsiepe, @thierrymoudiki

1. I have an inquiry in concerning verifying whether or not the license is acceptable.  I'll provide feedback on this ASAP.
2. The paper must compile.  @thierrymoudiki to bring clarity to this a bit for my sake, **please respond here** with the branch that represents the most current changes to the paper and to the package and supporting test suite.  We had set the active branch to the `paper` branch earlier; I want to make sure that is still the case.  Ultimately, we want to use a version of your code that has been reviewed and that matches what you have described in your paper.  If this is not the version currently on main, that is OK; but, we need a version that we can do an apples to apples comparison on for tests and what is in the paper.
3. @thierrymoudiki - please point me to a "how to contribute" section in your outward facing (docs or README) content.

I'll revisit this as soon as I hear back.
