### 2025-04-11 #6754 @crvernon
@spholmes, @bsiepe, @thierrymoudiki

I did just hear back rather quickly that if the license displayed in the repository and/or archive is not OSI approved (see https://opensource.org/licenses) then it is not acceptable by JOSS standards.  Therefore, a listed OSI approved license must be used if we are to continue.
