+++ vignettes/glmthetaf-ridge-attention.Rmd
@@ -1,0 +1,22 @@
+---
+title: "Ridge Theta Forecast with attention - conformal uncertainty"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Ridge Theta Forecast with attention - conformal uncertainty}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+# USAccDeaths
+
+
+```{r fig.width=7.5}
+(obj <- suppressWarnings(ahead::glmthetaf(USAccDeaths, h=25L, fit_func=MASS::glm.nb, attention = TRUE, type_pi = "conformal-split", method="adj")))
+plot(obj)
+```
+
+# AirPassengers
+
+```{r fig.width=5.5}  
+(obj <- suppressWarnings(ahead::glmthetaf(AirPassengers, h=25L, fit_func=MASS::glm.nb, attention = TRUE, type_pi = "conformal-split", method="adj")))
+```
