### 2025-04-11 #6754 @thierrymoudiki
> [@spholmes](https://github.com/spholmes), [@bsiepe](https://github.com/bsiepe), [@thierrymoudiki](https://github.com/thierrymoudiki)
>
> I did just hear back rather quickly that if the license displayed in the repository and/or archive is not OSI approved (see https://opensource.org/licenses) then it is not acceptable by JOSS standards. Therefore, a listed OSI approved license must be used if we are to continue.

Here's the license (same as BSD3 Clause): https://choosealicense.com/licenses/bsd-3-clause-clear/

Now, let's move on to something else, please. 5 pages -> 1 year: of course, everything moves in between. In my universe, this could have been wrapped up in a week (hopefully I'm not living there on my own, and sorry for the new reviewer @bsiepe  who may not understand my annoyance). Will update the branch within the next 30 minutes.
