### 2025-04-11 #6754 @crvernon
@thierrymoudiki - if the "clear" license cannot be found in https://opensource.org/licenses then we will not accept it.  Let me know if you are able to find the "clear" version of your license in the list provided and we will proceed to address the other issues.

Also, as a reminder, your submission is being edited and reviewed by a team of volunteers with time demands and priorities in their own respective careers.  These individuals donate their time to help authors like yourself have a forum to be recognized for the great work being in open source research software development.
