+++ vignettes/paper.Rmd
@@ -1,237 +1,0 @@
----
-title: 'ahead: Univariate and multivariate time series forecasting with uncertainty quantification including simulation approaches'
-tags:
-- Python
-- R
-- Julia
-- Time series
-- Forecasting
-- Machine Learning
-- Uncertainty
-date: "March 25, 2025"
-output:
-  pdf_document: default
-  word_document: default
-  html_document:
-    df_print: paged
-authors:
-- name: Thierry Moudiki
-  orcid: "0000-0002-9422-5459"
-  affiliation: 1
-affiliations:
-- name: "Techtonique LLC"
-  index: 1
-bibliography: biblio.bib
----
-
-# Summary
-
-`ahead` contains multiple time series forecasting models, and provides uncertainty quantification through parametric and non-parametric methods, including simulation-based approaches. This paper presents two original Machine Learning models implemented in R version of `ahead` package for forecasting univariate and multivariate time series: `dynrmf`, an autoregressive model that can utilize any Machine Learning model for forecasting univariate time series, and `ridge2f` that extends ridge regression with two regularization parameters and a hidden layer for producing nonlinear outputs.
-
-# Statement of need
-
-Forecasting multivariate time series (MTS hereafter, and including the univariate case) is important for business planning and decision support in finance (stress-testing with financial scenarios), insurance (reserving and required capital valuation), and other industries such as *Energy* (load anticipation) and meteorology. One can obtain point forecasts from a statistical/Machine Learning (ML) model, but these point forecasts are generally of limited importance to analysts. What matters more is the model's ability to quantify the uncertainty around its predictions. 
-
-There are multiple MTS forecasting models available in [\textsf{R} package](https://github.com/Techtonique/ahead) `ahead`'s version `0.18.0` ([\textsf{Python}](https://github.com/Techtonique/ahead_python) and [\textsf{Julia}](https://github.com/Techtonique/Ahead.jl) implementations, following \textsf{R}'s API as closely as possible). R version of `ahead` itself is available at  [r-packages.techtonique.net](https://r-packages.techtonique.net/), and distributed across all major operating systems.
-
-`ahead`'s models generally include parametric prediction intervals alongside non-parametric ones,  simulation-based uncertainty quantification techniques. This paper describes **two** of these ML models, **not available in any other statistical software**:
-
-* \mbox{\texttt{dynrmf}}; an autoregressive dynamic model inspired by **N**eural **N**etwork **A**uto**r**egression (NNAR) (from @hyndman2013forecasting), and described with more details in [https://otexts.com/fpp2/nnetar.html#neural-network-autoregression](https://otexts.com/fpp2/nnetar.html#neural-network-autoregression). As NNAR, \mbox{\texttt{dynrmf}} does an automatic choice of the number of autoregressive and seasonal time series lags. \mbox{\texttt{dynrmf}} is however more generic, and __can use any ML model__ (the default model is [Ridge Regression](https://en.wikipedia.org/wiki/Ridge_regression)).
-
[…]
