+++ NAMESPACE
@@ -42,6 +42,7 @@ export("ridge2f")
 export("ridge")
 export("varf")
 export("createtrendseason")
+export("get_error")
 export("getreturns")
 export("getsimulations")
 export("removenas")
