### 2025-04-11 #6754 @thierrymoudiki
@crvernon @spholmes  It's the same license. That's it. I can just remove that word if it really matters among all the other things.

I've just updated the branch with all 88 tests passing.

I always hear this excuse about _volunteers_. I have asked to volunteer in the past; no answer. Meaning, there are enough people (seemingly, but we know, right?). **Volunteering doesn't mean accept mediocrity**, and we should stop making it sound like that. You'll agree with me that this work doesn't require years. Okay, there are no reviewers, and some of them are simply **vanishing**. For 5 pages.

With that being said, I love the fact that my reviews are public these days (because this JOSS review is just a symptom of something bigger, see https://thierrymoudiki.github.io/blog/2025/01/20/r/forecasting/python/misc/ijf-benchmark-rejection-pt2), and everyone can appreciate the fact that **there's something else going on**, affecting everything else.

Now, if this takes more than 1 month, that is until May 11. Just REJECT. I'm done.
