### 2025-04-12 #6754 @arfon
@thierrymoudiki – your attitude here is way off the mark. As such, I've locked this thread for 7 days. You are invited to email [admin@theoj.org](mailto:admin@theoj.org) during this time to demonstrate a more collaborative attitude. Otherwise, we will proceed with rejection.
