+++ R/glmtheta.R
@@ -8,63 +8,169 @@
 #' @param x The time series data
 #' @param attention Logical flag for using attention mechanism
 #' @param scale_ctxt Scaling coefficient for context vector 
+#' @param B Number of bootstrap replications or number of simulations (yes, 'B' is unfortunate)
+#' @param nsim Alias for B
 #' @param ... Additional arguments to pass to the fit_func
 #' @return A forecast object
 #' @export
 glmthetaf <- function (y, 
                        h = ifelse(frequency(y) > 1, 2 * frequency(y), 10), 
-                       level = 95, 
+                       level = 95L, 
                        fit_func = stats::glm, 
                        fan = FALSE, 
                        x = y, 
                        type_pi = c(
+                         "conformal-surrogate",
+                         "conformal-kde",
+                         "conformal-bootstrap",
+                         "conformal-block-bootstrap",
+                         "conformal-fitdistr",
                          "conformal-split",
                          "gaussian"
                        ),
                        attention = TRUE, 
                        scale_ctxt = 1,
-                       method = c("adj", "base"),
+                       B = 250L,
+                       nsim = B,
                        ...) 
 {
-  method <- match.arg(method) 
   type_pi <- match.arg(type_pi)
   stopifnot(scale_ctxt > 0 && scale_ctxt <= 1)
+  method <- 
+  
   if (grepl("conformal", type_pi) >= 1) # not conformal
   {
     stopifnot(length(level) == 1)
     freq_x <- frequency(y)
     start_x <- start(y)
     start_preds <- tsp(y)[2] + 1 / freq_x
+    
     # Split the training data
     y_train_calibration <- splitts(y, split_prob=0.5)
     y_train <- y_train_calibration$training 
     y_calibration <- y_train_calibration$testing
     h_calibration <- length(y_calibration)
+    
     # Get predictions on calibration set
     y_pred_calibration <- ahead::glmthetaf(
       y_train,
       h = h_calibration,
       fit_func = fit_func, 
       attention = attention,
-      method = method,
-      type_pi = "gaussian"
+      type_pi = "gaussian",
+      ...
     )$mean
+    
     # Final fit and forecast on full calibration set
     fit_obj_train <- ahead::glmthetaf(
       y_calibration, 
       h = h,
       fit_func = fit_func,
       attention = attention,
-      method = method,
-      type_pi = "gaussian" 
+      type_pi = "gaussian",
+      ... 
     )
+    
+    fit_obj_train$method <- paste0("Conformal (", 
+                                  base::gsub("conformal-", "", type_pi), 
+                                  ") Theta")
+
     preds <- fit_obj_train$mean
     calibrated_residuals <- y_calibration - y_pred_calibration
-    scaled_calibrated_residuals <- base::scale(calibrated_residuals)
+    scaled_calib_resids <- base::scale(calibrated_residuals)
     sd_calibrated_residuals <- sd(calibrated_residuals)
+    
+    if (base::gsub("conformal-", "", type_pi) == "fitdistr")
+    {                            
+      simulate_function <- misc::fit_param_dist(as.numeric(scaled_calib_resids), 
+                                                verbose = FALSE)
+      sims <- matrix(simulate_function(nsim*h), ncol=nsim, nrow=h)
+      preds <- as.numeric(fit_obj_train$mean) + sims*sd_calibrated_residuals
+      
[…]
