+++ R/conformalize.R
@@ -24,7 +24,8 @@
 #' obj <- conformalize(FUN=HoltWinters, y=y, h=h, seasonal = "mult"); plot(obj)
 #' 
 conformalize <- function(FUN, y, h, level=95,
-                         method = c("block-bootstrap", "surrogate", "kde", "bootstrap"),
+                         method = c("block-bootstrap", "surrogate", 
+                                    "kde", "bootstrap", "fitdistr"),
                          nsim = 100L, 
                          block_size = 5,
                          seed = 123L, 
