### 2025-04-14 b71b002
add more conformalization methods Pt.2
