### 2025-04-20 #6754 @arfon
@thierrymoudiki – it is now a week since [my message to you](https://github.com/openjournals/joss-reviews/issues/6754#issuecomment-2799046827) inviting you to contact us privately to discuss how to proceed with this review. Having received no response, and therefore no indication that your approach to this review will change, we are now moving forward with a decision.

JOSS relies on the ability of all participants—authors, reviewers, and editors alike—to engage constructively and respectfully throughout the review process. This expectation is outlined in our [Code of Conduct](https://github.com/openjournals/joss/blob/main/CODE_OF_CONDUCT.md), which all authors agree to abide by upon submission. This includes, among other things, gracefully accepting constructive criticism and showing empathy toward other community members.

Your recent comments have demonstrated an inability to engage in this way, and having not heard from you in the past week, we can only conclude that you have no intention of changing your behavior.

As such, we are rejecting your submission to JOSS.

Arfon Smith, Editor in Chief.
